routes restructured into separate directories
