+++ routes/loader/animelist.js
