+++ routes/manga/loadlatest.js
@@ -1,7 +1,5 @@
 const express = require("express");
-const request = require("request");
-const cheerio = require("cheerio");
-const mangaPark = require("./mangapark");
+const mangaPark = require("../../utils/mangapark");
 const mangaParkObj = new mangaPark();
 const router = express.Router();
