+++ routes/search/search.js
@@ -1,7 +1,7 @@
 var express = require("express");
 const request = require("request");
 const cheerio = require("cheerio");
-const mangaPark = require("./mangapark");
+const mangaPark = require("../../utils/mangapark");
 const mangaParkObj = new mangaPark();
 
 var router = express.Router();
